+++ persistence/context.go
@@ -20,8 +20,14 @@ func (p PostgresContext) AppHash() ([]byte, error) {
 	return []byte("A real app hash, I am not"), nil
 }
 
-func (p PostgresContext) Reset() error {
-	panic("TODO: PostgresContext Reset not implemented")
+func (p *PostgresContext) Reset() error {
+	p.latestTxResults = nil
+	p.latestBlockHash = ""
+	p.latestQC = nil
+	p.latestProposerAddr = nil
+	p.latestBlockProtoBytes = nil
+	p.latestBlockTxs = nil
+	return nil
 }
 
 func (p PostgresContext) Commit() error {
@@ -29,6 +35,9 @@ func (p PostgresContext) Commit() error {
 
 	ctx := context.TODO()
 	if err := p.GetTx().Commit(context.TODO()); err != nil {
+		return err
+	}
+	if err := p.StoreBlock(); err != nil {
 		return err
 	}
 	if err := p.conn.Close(ctx); err != nil {
@@ -47,6 +56,9 @@ func (p PostgresContext) Release() error {
 	}
 	if err := p.conn.Close(ctx); err != nil {
 		log.Println("[TODO][ERROR] Implement connection pooling. Error when closing DB connecting...", err)
+	}
+	if err := p.Reset(); err != nil {
+		return err
 	}
 	return nil
 }